requirements.txt
black
pylint
requests
oauthlib
pyOpenSSL
flask<3
werkzeug<4
Flask-Login
python-dotenv
azure-storage-file-datalake
azure-identity
pandas
chardet